+++ src/polymer/polymer2-mixin-scanner.ts
@@ -14,7 +14,7 @@
 
 import * as estree from 'estree';
 
-import {getIdentifierName} from '../javascript/ast-value';
+import {getIdentifierName, getNamespacedIdentifier} from '../javascript/ast-value';
 import {Visitor} from '../javascript/estree-visitor';
 import * as esutil from '../javascript/esutil';
 import {JavaScriptDocument} from '../javascript/javascript-document';
@@ -51,15 +51,18 @@ class MixinVisitor implements Visitor {
     if (parent.type !== 'ExpressionStatement') {
       return;
     }
-    const comment = esutil.getAttachedComment(parent) || '';
-    const docs = jsdoc.parseJsdoc(comment);
-    if (this._hasPolymerMixinDocTag(docs)) {
+    const parentComments = esutil.getAttachedComment(parent) || '';
+    const parentJsDocs = jsdoc.parseJsdoc(parentComments);
+    if (this._hasPolymerMixinDocTag(parentJsDocs)) {
       const name = getIdentifierName(node.left);
+      const namespacedName =
+          name ? getNamespacedIdentifier(name, parentJsDocs) : undefined;
       const sourceRange = this._document.sourceRangeForNode(node);
       this._currentMixin = new ScannedPolymerElementMixin({
-          name, sourceRange,
-          // TODO(justinfagnani): fix descriptions correctly in parseJsdoc
-          // description: docs.description,
+        name: namespacedName,
+        sourceRange,
+        // TODO(justinfagnani): fix descriptions correctly in parseJsdoc
+        // description: docs.description,
       });
       this._currentMixinNode = node;
       this._mixins.push(this._currentMixin);
@@ -68,16 +71,19 @@ class MixinVisitor implements Visitor {
 
   enterFunctionDeclaration(
       node: estree.FunctionDeclaration, _parent: estree.Node) {
-    const comment = esutil.getAttachedComment(node) || '';
-    const docs = jsdoc.parseJsdoc(comment);
-    if (this._hasPolymerMixinDocTag(docs)) {
+    const nodeComments = esutil.getAttachedComment(node) || '';
+    const nodeJsDocs = jsdoc.parseJsdoc(nodeComments);
+    if (this._hasPolymerMixinDocTag(nodeJsDocs)) {
       const name = node.id.name;
+      const namespacedName =
+          name ? getNamespacedIdentifier(name, nodeJsDocs) : undefined;
       const sourceRange = this._document.sourceRangeForNode(node);
       this._currentMixinFunction = node;
       this._currentMixin = new ScannedPolymerElementMixin({
-          name, sourceRange,
-          // TODO(justinfagnani): fix descriptions correctly in parseJsdoc
-          // description: docs.description,
+        name: namespacedName,
+        sourceRange,
+        // TODO(justinfagnani): fix descriptions correctly in parseJsdoc
+        // description: docs.description,
       });
       this._currentMixinNode = node;
       this._mixins.push(this._currentMixin);
@@ -120,9 +126,12 @@ class MixinVisitor implements Visitor {
   }
 
   enterVariableDeclarator(
-      node: estree.VariableDeclarator, _parent: estree.Node) {
+      node: estree.VariableDeclarator, parent: estree.Node) {
     if (this._currentMixin != null && this._currentMixinFunction == null) {
-      this._currentMixin.name = (node.id as estree.Identifier).name;
+      const name = (node.id as estree.Identifier).name;
+      const parentComments = esutil.getAttachedComment(parent) || '';
+      const parentJsDocs = jsdoc.parseJsdoc(parentComments);
+      this._currentMixin.name = getNamespacedIdentifier(name, parentJsDocs);
     }
   }