+++ src/test/polymer/polymer2-mixin-scanner_test.ts
@@ -103,7 +103,7 @@ function TestMixin(superclass) {
     const mixins = await getMixins('test-mixin-2.js');
     const mixinData = mixins.map(getTestProps);
     assert.deepEqual(mixinData, [{
-                       name: 'TestMixin',
+                       name: 'Polymer.TestMixin',
                        description: '',
                        properties: [{
                          name: 'foo',
@@ -140,7 +140,7 @@ const TestMixin = (superclass) => class extends superclass {
     const mixins = await getMixins('test-mixin-3.js');
     const mixinData = mixins.map(getTestProps);
     assert.deepEqual(mixinData, [{
-                       name: 'TestMixin',
+                       name: 'Polymer.TestMixin',
                        description: '',
                        properties: [{
                          name: 'foo',
@@ -183,7 +183,7 @@ const TestMixin = function(superclass) {
         const mixins = await getMixins('test-mixin-4.js');
         const mixinData = mixins.map(getTestProps);
         assert.deepEqual(mixinData, [{
-                           name: 'TestMixin',
+                           name: 'Polymer.TestMixin',
                            description: '',
                            properties: [],
                            attributes: [],
@@ -205,7 +205,7 @@ let TestMixin;
     const mixins = await getMixins('test-mixin-6.js');
     const mixinData = mixins.map(getTestProps);
     assert.deepEqual(mixinData, [{
-                       name: 'TestMixin',
+                       name: 'Polymer.TestMixin',
                        description: '',
                        properties: [],
                        attributes: [],
@@ -262,5 +262,23 @@ Polymer.TestMixin = Polymer.woohoo(function TestMixin(base) {
 });
 ~~`);
   });
+
+  test(
+      'properly analyzes nested mixin assignments with memberof tags',
+      async() => {
+        const mixins = await getMixins('test-mixin-8.js');
+        const mixinData = mixins.map(getTestProps);
+        assert.deepEqual(mixinData, [{
+                           name: 'Polymer.TestMixin',
+                           description: '',
+                           properties: [{
+                             name: 'foo',
+                           }],
+                           attributes: [{
+                             name: 'foo',
+                           }],
+                         }]);
+
+      });
 
 });