+++ src/test/static/js-behaviors.js
@@ -3,8 +3,20 @@ var SimpleBehavior = {
   simple: true,
 };
 
+/**
+ * @polymerBehavior
+ * @memberof Polymer
+ * */
+var SimpleNamespacedBehavior = {
+  simple: true,
+};
+
+
 /** @polymerBehavior AwesomeBehavior */
 var CustomNamedBehavior = {custom: true, properties: {a: {value: 1}}};
+
+/** @polymerBehavior Polymer.AwesomeNamespacedBehavior */
+var CustomNamedBehaviorTwo = {custom: true, properties: {a: {value: 1}}};
 
 /**
 With a chained behavior