+++ src/test/static/polymer2/test-mixin-3.js
@@ -1,5 +1,6 @@
 /**
  * @polymerMixin
+ * @memberof Polymer
  */
 const TestMixin = function(superclass) {
   return class extends superclass {