+++ src/test/static/polymer2/test-mixin-4.js
@@ -1,5 +1,6 @@
 /**
  * @polymerMixin
+ * @memberof Polymer
  */
 let TestMixin;