+++ src/test/static/polymer2/test-mixin-5.js
@@ -1,5 +1,6 @@
 /**
  * @polymerMixin
+ * @memberof Polymer
  */
 class TestMixin extends superclass {
   static get properties() {