+++ src/test/static/polymer2/test-mixin-6.js
@@ -1,5 +1,6 @@
 /**
  * @polymerMixin
+ * @memberof Polymer
  */
 function TestMixin() {
 }